moved relative coordinates to coordinates submodule
