+++ dynhalo/finder/coordinates.py
@@ -1,0 +1,38 @@
+import numpy as np
+
+
+def relative_coordinates(
+    x0: np.ndarray,
+    x: np.ndarray,
+    boxsize: float,
+    periodic: bool = True
+) -> float:
+    """Returns the coordinates x relative to x0 accounting for periodic boundary
+    conditions
+
+    Parameters
+    ----------
+    x0 : np.ndarray
+        Reference position in cartesian coordinates
+    x : np.ndarray
+        Position array (N, 3)
+    boxsize : float
+        Size of simulation box
+    periodic : bool, optional
+        Set to True if the simulation box is periodic, by default True
+
+    Returns
+    -------
+    float
+        Relative positions
+    """
+    if periodic:
+        return (x - x0 + 0.5*boxsize) % boxsize - 0.5*boxsize
+    return x - x0
+
+
+def rec_to_sph_coordinates(rec: np.ndarray):
+    return
+
+if __name__ == "__main__":
+    pass
