+++ dynhalo/finder/halo.py
@@ -3,36 +3,6 @@ from typing import Tuple
 import numpy as np
 
 
-def relative_coordinates(
-    x0: np.ndarray,
-    x: np.ndarray,
-    boxsize: float,
-    periodic: bool = True
-) -> float:
-    """Returns the coordinates x relative to x0 accounting for periodic boundary
-    conditions
-
-    Parameters
-    ----------
-    x0 : np.ndarray
-        Reference position in cartesian coordinates
-    x : np.ndarray
-        Position array (N, 3)
-    boxsize : float
-        Size of simulation box
-    periodic : bool, optional
-        Set to True if the simulation box is periodic, by default True
-
-    Returns
-    -------
-    float
-        Relative positions
-    """
-    if periodic:
-        return (x - x0 + 0.5*boxsize) % boxsize - 0.5*boxsize
-    return x - x0
-
-
 def find_r200_m200(
     r_rel: np.ndarray,
     part_mass: float,
